Most recent version.
This is actually a combination of my ADnD Framework, based upon Celestian's, and my Greyhawk Library. Nothing should be too broken, but it's in quite a shambles. We are making changes as we go along, and occasionally find that a change has broken something. On the other hand, we're also fixing things with the old framework that broke when we upgraded to v 1.8.4 of MapTool.

Feel free to submit issues, and I'll hopefully get them fixed in time. I'm starting to play with Foundry (having shelled out for it) and may eventually do my development over there. Although the systems are completely different, the logic will probably be at least somewhat similar, and I may be able to port some of the stuff from there to MT macros. And then there's always the promise of 2.0.